requirements.txt
Flask==3.0.3 
google-generativeai==0.8.2 
reportlab==4.2.2